Improve ensemble client and add new tests
