+++ tests/ensemble_evaluator/conftest.py
@@ -2,7 +2,8 @@ import json
 import os
 from pathlib import Path
 from unittest.mock import Mock
-
+import asyncio
+import websockets
 import pytest
 from ert_shared.ensemble_evaluator.entity.ensemble import (
     create_ensemble_builder,
@@ -139,3 +140,24 @@ def _dump_ext_job(ext_job, index):
         "arg_types": ext_job.arg_types,
         "argList": ext_job.get_arglist(),
     }
+
+
+def _mock_ws(host, port, messages, delay_startup=0):
+    loop = asyncio.new_event_loop()
+    done = loop.create_future()
+
+    async def _handler(websocket, path):
+        while True:
+            msg = await websocket.recv()
+            messages.append(msg)
+            if msg == "stop":
+                done.set_result(None)
+                break
+
+    async def _run_server():
+        await asyncio.sleep(delay_startup)
+        async with websockets.serve(_handler, host, port):
+            await done
+
+    loop.run_until_complete(_run_server())
+    loop.close()
